+++ lib/themed.dart
@@ -1,1 +1,2 @@
+export 'src/color_util.dart';
 export 'src/themed.dart';